feat. bookSeats 관련 기능 구현

+++ src/main/java/com/concert_reservation/api/application/dto/response/UserResponse.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class UserResponse {
     private String userId;
     private String name;
-    private int balance;
+    private Long balance;
 }
